Began solving #196 because of that we need to implement the Properties of the PasswordManagerItem as much as possible as Arrays and not as Strings because a String can't be overriden in a save context and will be way harder to GC Collect as an Array.
For this approach I implemented additional extensions for the char[] to have a similar experience likely to a string
I also changed other Properties storing type from a string to an int and from a string to an enum.

We also need to implement for each DatabaseFile Version it's own DatabasePasswordItem since it will be not possible to load the new PasswordMangerItems with old Versions because they get stored diffrent.

Because of that we need always to have a base store, which we need to convert to, to have a "Standard".

To do:
- We need to implement a Logic to override everything after a logout
- We need to finish implementing the DatabasePasswordItem's to be able to load Databases
- We need to override the String (which we need) after it's use for example if we want to show the string we convert it from a SecureString or Array depends on what it will be later to a String. If we are done to show the string we should zero it out to be sure it won't appear in the memory with the password as value